requirements.txt
python-telegram-bot==20.3
Pillow==10.0.0
python-dotenv==1.0.0
peewee==3.17.0